% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_data.frame.R, R/as_data.table.R, R/as_ts.R,
%   R/as_xts.R
\name{as_tbl}
\alias{as_tbl}
\alias{as_data.frame}
\alias{as_df}
\alias{as_data.frame.xts}
\alias{as_data.frame.ts}
\alias{as_data.frame.data.frame}
\alias{as_data.frame.data.table}
\alias{as_data.frame.tbl}
\alias{as_data.table}
\alias{as_dt}
\alias{as_ts}
\alias{as_ts.xts}
\alias{as_ts.data.frame}
\alias{as_ts.data.table}
\alias{as_ts.ts}
\alias{as_xts}
\alias{as_xts.ts}
\alias{as_xts.xts}
\alias{as_xts.data.frame}
\alias{as_xts.data.table}
\title{Convert everything to everything}
\usage{
as_tbl(x, ...)

as_data.frame(x, ...)

as_df(x, ...)

\method{as_data.frame}{xts}(x, time.name = getOption("tsbox.time.name",
  "time"), var.name = getOption("tsbox.var.name", "var"),
  value.name = getOption("tsbox.value.name", "value"), ...)

\method{as_data.frame}{ts}(x, ...)

\method{as_data.frame}{data.frame}(x, ...)

\method{as_data.frame}{data.table}(x, ...)

\method{as_data.frame}{tbl}(x, ...)

as_data.table(x, ...)

as_dt(x, ...)

as_ts(x, ...)

\method{as_ts}{xts}(x, ...)

\method{as_ts}{data.frame}(x, ...)

\method{as_ts}{data.table}(x, ...)

\method{as_ts}{ts}(x, ...)

as_xts(x, ...)

\method{as_xts}{ts}(x, ...)

\method{as_xts}{xts}(x, ...)

\method{as_xts}{data.frame}(x, time.name = getOption("tsbox.time.name",
  "time"), var.name = getOption("tsbox.var.name", "var"),
  value.name = getOption("tsbox.value.name", "value"), ...)

\method{as_xts}{data.table}(x, time.name = getOption("tsbox.time.name",
  "time"), var.name = getOption("tsbox.var.name", "var"),
  value.name = getOption("tsbox.value.name", "value"), ...)
}
\arguments{
\item{x}{a time series object, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}

\item{...}{additional arguments, passed to methods}

\item{time.name}{time name (default: \code{"time"})}

\item{var.name}{var name (default: \code{"var"})}

\item{value.name}{value name (default: \code{"time"})}
}
\description{
Convert everything to everything
}
\examples{

# optionally changes column names (data.frame and data.table) methods
options(tsbox.time.name = "Date",
        tsbox.value.name = "Value",
        tsbox.var.name = "Variable"
        )

x.ts <- tsbind(mdeaths, fdeaths) 
x.xts <- as_xts(x.ts)
x.df <- as_df(x.xts)
\dontrun{
library(data.table)
x.dt <- as_dt(x.df) 
}

}
