% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coerce_to_}
\alias{coerce_to_}
\title{Universal Converter Function}
\usage{
coerce_to_(x = "xts")
}
\arguments{
\item{x}{time series object, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}
}
\value{
returns a function
}
\description{
Universal Converter Function
}
