% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_from_all.R
\name{copy_class}
\alias{copy_class}
\title{Reclass Time Series}
\usage{
copy_class(z, x)
}
\arguments{
\item{z}{series to reclass}

\item{x}{template series}
}
\description{
Copies class attributes from an existing ts-boxable series. Mainly used
internally.
}
\details{
Inspired by \code{xts::reclass}, which does something similar.
}
