% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{date_shift}
\alias{date_shift}
\alias{date_shift}
\alias{date_month}
\alias{date_shift}
\alias{date_quarter}
\alias{date_shift}
\alias{date_year}
\title{Manipulating Dates}
\usage{
date_shift(x, by = NULL)

date_month(x)

date_quarter(x)

date_year(x)
}
\arguments{
\item{x}{\code{Date} or \code{POSIXct}. If \code{POSIXct}, it is converted into \code{Date}.}

\item{by}{integer or character, either the number of shifting periods
(integer), or an absolute amount of time (character). See details.}
}
\value{
an object of class \code{Date}
}
\description{
\code{date_shift} adds seconds, minutes, hours, days, weeks, months, quarters or years to dates.
\code{date_month}, \code{date_quarter} and \code{date_yeae} return the first day of the
period and are useful for customized aggregation of data frames. For standard
aggregation, use \code{\link[=ts_frequency]{ts_frequency()}}.
}
\details{
If \code{by} is character, the time stamp is shifted by a specific amount of time.
This can be one of one of \code{"sec"}, \code{"min"}, \code{"hour"}, \code{"day"}, \code{"week"},
\code{"month"}, \code{"quarter" or}"year", optionally preceded by a (positive or
negative) integer and a space, or followed by plural "s". This is passed to
\code{\link[base:seq.Date]{base::seq.Date()}}. This does not require the series to be regular.
}
\examples{
ap.time <- ts_df(AirPassengers)$time
head(date_month(ap.time))
head(date_quarter(ap.time))
head(date_year(ap.time))

head(date_shift(ap.time, 14))
head(date_shift(ap.time, "7 week"))
head(date_shift(ap.time, "-1 month"))

date_shift(ts_end(mdeaths), 1)
date_shift(ts_end(mdeaths), "-14 sec")
date_shift(ts_end(mdeaths), "-1 year")
ts_span(ts_c(mdeaths, fdeaths), start = date_shift(ts_end(mdeaths), -1))

}
\seealso{
\code{\link[=ts_frequency]{ts_frequency()}} for standard aggregation. \code{\link[=date_shift]{date_shift()}}, for
shifting time stamps of a ts-boxable object.
}
