% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_properties.R
\name{from}
\alias{from}
\alias{to}
\alias{freq}
\alias{ids}
\title{Extract Time Series Properties}
\usage{
from(x)

to(x)

freq(x, type = c("freq", "string"))

ids(x, simplify = TRUE)
}
\arguments{
\item{x}{ts-boxable time series, an object of class \code{ts}, \code{xts}, \code{zoo}, \code{data.frame}, \code{data.table}, \code{tbl}, \code{tbl_ts}, \code{tbl_time}, or \code{timeSeries}.}

\item{type}{decimal or by string}
}
\description{
These functions do not start with \code{ts_}, because they don't return an object
of the same class as the input.
}
