% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_from_all.R
\name{relevant_class}
\alias{relevant_class}
\title{Extract the Relavant Class}
\usage{
relevant_class(x)
}
\arguments{
\item{x}{time series object, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}
}
\description{
Extract the Relavant Class
}
