% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_append.R
\name{ts_append}
\alias{ts_append}
\alias{ts_append}
\alias{ts_append.ts}
\alias{ts_append}
\alias{ts_append.data.frame}
\alias{ts_append}
\alias{ts_append.xts}
\alias{ts_append}
\alias{ts_append.zoo}
\alias{ts_append}
\alias{ts_append.data.table}
\title{Append values to a time series}
\usage{
ts_append(series, v)

\method{ts_append}{ts}(series, v)

\method{ts_append}{data.frame}(series, v)

\method{ts_append}{xts}(series, v)

\method{ts_append}{zoo}(series, v)

\method{ts_append}{data.table}(series, v)
}
\arguments{
\item{series}{object of class ts, xts, zoo, data.frame or data.table
representing a time series.}

\item{v}{vector to be appended at the end of the existing series.}
}
\description{
Add a one or more elements of a vector to the end of a time series.
This is an S3 implementation such that appending works for a ts, xts, zoo,
data.frame and data.table base.
}
\examples{
ts1 <- ts(1:20,start = c(2000,2),frequency = 4)
ts1_df <- ts_df(ts1)
ts1_dt <- ts_data.table(ts1)
# Different append S3 methods
ts_append(ts1,30)
ts_append(ts1_df,c(30,40))
ts_append(ts1_dt,100)
}
\author{
Matthias Bannert
}
