% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_apply.R
\name{ts_apply}
\alias{ts_apply}
\title{Apply Function on Multiple Time Series}
\usage{
ts_apply(x, FUN, ...)
}
\arguments{
\item{x}{time series objects, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}

\item{FUN}{a function that can applied on the corresponding object}

\item{...}{additional arguments, passed to FUN}
}
\description{
Apply Function on Multiple Time Series
}
