% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_bind.R, R/ts_chain.R
\name{ts_bind}
\alias{ts_bind}
\alias{ts_bind}
\alias{ts_chain}
\title{Combine to Single Time Series}
\usage{
ts_bind(...)

ts_chain(...)
}
\arguments{
\item{...}{one or several tsboxable time series}
}
\description{
Combine time series to a new, single time series. \code{ts_bind} combines time
series as they are, \code{ts_chain} chaines them together, using percentage change
rates.
}
\examples{
ts_bind(ts_window(mdeaths, end = "1975-12-01"), fdeaths)
ts_bind(mdeaths, c(2, 2))
ts_bind(mdeaths, 3, ts_bind(fdeaths, c(99, 2)))
ts_bind(ts_dt(mdeaths), AirPassengers)
ts_chain(ts_window(mdeaths, end = "1975-12-01"), fdeaths)
ts_plot(ts_pc(ts_c(
  comb = ts_chain(ts_window(mdeaths, end = "1975-12-01"), fdeaths), 
  fdeaths
)))
}
