% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_from_all.R
\name{ts_boxable}
\alias{ts_boxable}
\title{Test if an Object is ts-Boxable}
\usage{
ts_boxable(x)
}
\arguments{
\item{x}{time series object, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}
}
\value{
logical, either \code{TRUE} or \code{FALSE}
}
\description{
Mainly used internally.
}
\examples{
ts_boxable(AirPassengers)
ts_boxable(lm)
}
