% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_c.R, R/ts_chain.R, R/ts_rbind.R
\name{ts_c}
\alias{ts_c}
\alias{ts_chain}
\alias{ts_chain}
\alias{ts_bind}
\title{Bind any time series vertically or horizontally}
\usage{
ts_c(...)

ts_chain(...)

ts_chain(...)

ts_bind(...)
}
\arguments{
\item{...}{time series objects, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}
}
\description{
Bind any time series vertically or horizontally

Bind to or several tsboxable objects, chaining them together where the first
series ends.

Bind to or several tsboxable objects
}
\examples{

ts_c(ts_df(EuStockMarkets), AirPassengers)
a <- ts_c(EuStockMarkets, mdeaths)

# labelling:
ts_c(`International Airline Passengers` = ts_xts(AirPassengers), 
       `Deaths from Lung Diseases` = ldeaths)

ts_c(a = mdeaths, AirPassengers)
ts_bind(ts_xts(AirPassengers), mdeaths)
ts_c(ts_dt(EuStockMarkets), AirPassengers)
ts_bind(ts_dt(mdeaths), AirPassengers)

}
