% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_separate.R
\name{ts_combine}
\alias{ts_combine}
\alias{ts_combine}
\alias{ts_separate}
\title{Combine or separate variable names in data.frame like objects}
\usage{
ts_combine(x)

ts_separate(x, into)
}
\arguments{
\item{x}{any time series object}

\item{into}{new column names}
}
\description{
Combine or separate variable names in data.frame like objects
}
