% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_complete.R
\name{ts_complete}
\alias{ts_complete}
\title{Make all time series the same length}
\usage{
ts_complete(x, fill = NA)
}
\arguments{
\item{x}{any time series object}

\item{fill}{missign value specifier}
}
\description{
Make all time series the same length
}
