% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts.R
\name{ts_dts}
\alias{ts_dts}
\title{Convert everything to everything}
\usage{
ts_dts(x, ...)
}
\arguments{
\item{x}{a time series object, either \code{ts}, \code{data.frame}, \code{data.table}, \code{tibble} or \code{xts}.}

\item{...}{additional arguments, passed to methods}

\item{time.name}{time name (default: \code{"time"})}

\item{var.name}{var name (default: \code{"var"})}

\item{value.name}{value name (default: \code{"time"})}
}
\description{
Convert everything to everything
}
\examples{

x.ts <- ts_bind(mdeaths, fdeaths) 
x.df <- ts_df(x.xts)
x.dt <- ts_dt(x.xts)
\dontrun{
library(xts)
x.xts <- ts_xts(x.ts)
library(dplyr)
x.tbl <- ts_tbl(x.ts)
}

}
