% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_forecast.R
\name{ts_forecast_mean}
\alias{ts_forecast_mean}
\alias{ts_forecast_auto.arima_mean}
\alias{ts_forecast_mean}
\title{Agnostic wrapper around forecast}
\usage{
ts_forecast_auto.arima_mean(x, xreg = NULL, h = 2, ...)

ts_forecast_mean(x, h = 2, ...)
}
\arguments{
\item{x}{any time series object}

\item{xreg}{any time series object}

\item{h}{forecast horizon}

\item{...}{arguments passed to main function}
}
\description{
Agnostic wrapper around forecast
}
