% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_frequency.R
\name{ts_frequency}
\alias{ts_frequency}
\alias{ts_frequency}
\alias{last}
\alias{ts_frequency}
\alias{first}
\title{Change Frequency}
\usage{
ts_frequency(x, to = "year", fun = mean)

last(x)

first(x)
}
\arguments{
\item{x}{a ts-boxable time series}

\item{to}{desired frequency, either a character string (\code{"year"},
\code{"quarter"}, \code{"month"}) or an integer (\code{1}, \code{4}, \code{12}).}

\item{fun}{aggregation function (\code{mean}, \code{sum}, \code{first}, \code{last})}
}
\description{
Changes the frequency of a ts boxable time series. Currently, incomplete
periods are aggregated as well, but this is likely to change.
}
\examples{
ts_frequency(cbind(mdeaths, fdeaths), "year", sum)
ts_frequency(cbind(mdeaths, fdeaths), "quarter", sum)

ts_frequency(AirPassengers, 4, sum)
ts_frequency(AirPassengers, 1, sum)

# Note that incomplete years are (currently) aggregated as well
ts_frequency(EuStockMarkets, "year")

}
