% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_reshape.R
\name{ts_gather}
\alias{ts_gather}
\alias{ts_gather.ts}
\alias{ts_gather.xts}
\alias{ts_gather.data.frame}
\alias{ts_gather.data.table}
\alias{ts_gather.tbl}
\alias{ts_spread.ts}
\alias{ts_spread.xts}
\alias{ts_spread.data.frame}
\alias{ts_spread.data.table}
\alias{ts_spread.tbl}
\title{Reshaping Multiple Time Series}
\usage{
ts_gather(x, ...)

\method{ts_gather}{ts}(x, ...)

\method{ts_gather}{xts}(x, ...)

\method{ts_gather}{data.frame}(x, ...)

\method{ts_gather}{data.table}(x, ...)

\method{ts_gather}{tbl}(x, ...)

\method{ts_spread}{ts}(x, ...)

\method{ts_spread}{xts}(x, ...)

\method{ts_spread}{data.frame}(x, ...)

\method{ts_spread}{data.table}(x, ...)

\method{ts_spread}{tbl}(x, ...)
}
\description{
Reshaping Multiple Time Series
}
