% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_ggplot.R
\name{ts_ggplot}
\alias{ts_ggplot}
\title{Beautiful Graphics}
\usage{
ts_ggplot(x, title = NULL, subtitle = NULL, ...)
}
\description{
Beautiful Graphics
}
\examples{
library(mytools)
ts_ggplot(AirPassengers) + ggtitle("Airline passengers", subtitle = "The classic Box & Jenkins airline data")
ts_ggplot(cbind(total = ldeaths, female = fdeaths, male = mdeaths))

library(Quandl)
ts_ggplot(Quandl("FRED/GDPMC1", "xts"))
ggsave("myfig.pdf", width = 8, height = 5)
}
