% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_pc.R
\name{ts_lag}
\alias{ts_lag}
\alias{ts_lag}
\alias{ts_pc}
\alias{ts_lag}
\alias{ts_diff}
\alias{ts_lag}
\alias{ts_index}
\alias{ts_lag}
\alias{ts_pcy}
\alias{ts_lag}
\alias{ts_diffy}
\title{Lags, Differences, Percentage Change Rates}
\usage{
ts_lag(x, lag = 1, fill = NA)

ts_pc(x)

ts_diff(x)

ts_index(x, base)

ts_pcy(x, ...)

ts_diffy(x, ...)
}
\arguments{
\item{x}{ts-boxable time series, an object of class \code{ts}, \code{xts}, \code{data.frame}, \code{data.table}, or \code{tibble}.}

\item{lag}{integer, the number of lags. This is defined as in dplyr, opposite to R base.}

\item{fill}{how to fill missing values}

\item{base}{base date, character string, \code{Date} or \code{POSIXct}, at which the
value will be set to 1.}

\item{...}{additional arguments, passed to subfunctions (as some of these functions are generated by \link{ts_})}
}
\value{
a ts-boxable time series, with the same class as the input.
}
\description{
Utlility functions to work with time series. \code{ts_pcy} and \code{ts_diffy} calulate
the percentage change and the difference compared to the same period of the
previous year, rather than the previous period. \code{ts_index} returns an index
series, with value of 1 at \code{base} date.
}
\examples{
ts_lag(ts_c(fdeaths, mdeaths))
ts_diff(ts_c(fdeaths, mdeaths))
ts_pc(ts_c(fdeaths, mdeaths))
ts_index(ts_df(ts_c(fdeaths, mdeaths)), "1977-01-01")
}
