% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_misc.R
\name{ts_misc}
\alias{ts_misc}
\alias{ts_prcomp}
\alias{ts_misc}
\alias{ts_dygraphs}
\alias{ts_misc}
\alias{ts_forecast}
\alias{ts_misc}
\alias{ts_seas}
\title{Useful functions, generated by \code{ts_}}
\usage{
ts_prcomp(x, ...)

ts_dygraphs(x, ...)

ts_forecast(x, ...)

ts_seas(x, ...)
}
\arguments{
\item{x}{a ts-boxable object}

\item{...}{further arguments, passed to functions}
}
\description{
Useful functions, generated by \code{ts_}
}
\examples{
ts_prcomp(ts_c(mdeaths, fdeaths))
ts_dygraphs(ts_c(mdeaths, EuStockMarkets))
ts_forecast(ts_c(mdeaths, fdeaths))
ts_seas(ts_c(mdeaths, fdeaths))

}
