% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsnames.R
\name{ts_names}
\alias{ts_names}
\alias{ts_names.ts}
\alias{ts_names.xts}
\alias{ts_names.data.frame}
\alias{ts_names.data.table}
\alias{ts_names.tbl}
\alias{ts_set_names}
\alias{ts_set_names.ts}
\alias{ts_set_names.xts}
\alias{ts_set_names.data.frame}
\alias{ts_set_names.data.table}
\alias{ts_set_names.tbl}
\title{Set and Extract Names of Time Series}
\usage{
ts_names(x, ...)

\method{ts_names}{ts}(x, ...)

\method{ts_names}{xts}(x, ...)

\method{ts_names}{data.frame}(x, ...)

\method{ts_names}{data.table}(x, ...)

\method{ts_names}{tbl}(x, ...)

ts_set_names(x, value)

\method{ts_set_names}{ts}(x, value)

\method{ts_set_names}{xts}(x, value)

\method{ts_set_names}{data.frame}(x, value)

\method{ts_set_names}{data.table}(x, value)

\method{ts_set_names}{tbl}(x, value)
}
\arguments{
\item{x}{a time series object, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}

\item{...}{additional arguments, passed to methods}

\item{value}{new names}
}
\description{
Set and Extract Names of Time Series
}
\examples{

x.ts <- ts_bind(mdeaths, fdeaths) 
x.xts <- ts_xts(x.ts)
x.df <- ts_df(x.xts)

ts_names(x.ts)
ts_names(x.xts)
ts_names(x.df)

\dontrun{
library(data.table)  # if you want to use the 'data.table' methods
x.dt <- ts_dt(x.df)
ts_names(x.dt)
}

}
