% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tspc.R, R/tstrend.R, R/tswin.R
\name{ts_pc}
\alias{ts_pc}
\alias{ts_pc.ts}
\alias{ts_pc.xts}
\alias{ts_pc.data.frame}
\alias{ts_pc.data.table}
\alias{ts_pc.tbl}
\alias{ts_pcy}
\alias{ts_pcy.ts}
\alias{ts_pcy.xts}
\alias{ts_pcy.data.frame}
\alias{ts_pcy.data.table}
\alias{ts_pcy.tbl}
\alias{ts_trend}
\alias{ts_window}
\alias{ts_window.xts}
\alias{ts_window.ts}
\alias{ts_window.data.frame}
\alias{ts_window.data.table}
\alias{ts_window.tbl}
\title{Utility Functions for Time Series}
\usage{
ts_pc(x, ...)

\method{ts_pc}{ts}(x, ...)

\method{ts_pc}{xts}(x, ...)

\method{ts_pc}{data.frame}(x, ...)

\method{ts_pc}{data.table}(x, ...)

\method{ts_pc}{tbl}(x, ...)

ts_pcy(x, ...)

\method{ts_pcy}{ts}(x, ...)

\method{ts_pcy}{xts}(x, ...)

\method{ts_pcy}{data.frame}(x, ...)

\method{ts_pcy}{data.table}(x, ...)

\method{ts_pcy}{tbl}(x, ...)

ts_trend(x, ...)

ts_window(x, ...)

\method{ts_window}{xts}(x, start = NULL, end = NULL, ...)

\method{ts_window}{ts}(x, ...)

\method{ts_window}{data.frame}(x, ...)

\method{ts_window}{data.table}(x, ...)

\method{ts_window}{tbl}(x, ...)
}
\arguments{
\item{x}{a time series object, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}

\item{...}{additional arguments, passed to methods}

\item{start}{start date, string, Date or POSIXct}

\item{end}{end date, string, Date or POSIXct}
}
\description{
Utility Functions for Time Series
}
\examples{
library(tsbox)

x.ts <- ts_bind(mdeaths, fdeaths) 
x.xts <- ts_xts(x.ts)
x.df <- ts_df(x.xts)

ts_scale(x.ts)
ts_scale(x.xts)
ts_scale(x.df)
ts_trend(x.ts)
ts_trend(x.xts)
ts_trend(x.df)

ts_pc(x.ts)
ts_pc(x.xts)
ts_pc(x.df)

ts_pcy(x.ts)
ts_pcy(x.xts)
ts_pcy(x.df)

\dontrun{
library(data.table)  # if you want to use the 'data.table' methods
x.dt <- ts_dt(x.df)
ts_scale(x.dt)
ts_trend(x.dt)
ts_pc(x.dt)
ts_pcy(x.dt)
}

}
