% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_properties.R
\name{ts_prop}
\alias{ts_prop}
\title{Time Series Properties}
\usage{
ts_prop(x, spark.width = 15)
}
\arguments{
\item{x}{ts-boxable time series, an object of class \code{ts}, \code{xts}, \code{zoo}, \code{data.frame}, \code{data.table}, \code{tbl}, \code{tbl_ts}, \code{tbl_time}, or \code{timeSeries}.}

\item{spark.width}{integer, with of the spark graph, in numer of character.
Set to \code{NULL} to turn off.}
}
\description{
Time Series Properties
}
