% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{ts_reclass}
\alias{ts_reclass}
\title{Reclass an object to a ts-boxable series}
\usage{
ts_reclass(z, x)
}
\arguments{
\item{z}{series to reclass}

\item{x}{template series}
}
\description{
Inspired by the similarly named function from the xts package
}
