% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{ts_save}
\alias{ts_save}
\title{Save Previous Plot}
\usage{
ts_save(filename = "myfig.pdf", width = 10, height = 5, device = "pdf",
  ..., open = TRUE)
}
\arguments{
\item{filename}{filename}

\item{width}{width}

\item{height}{height}

\item{device}{device}

\item{...}{additional arguments}

\item{open}{open}
}
\description{
Save Previous Plot
}
