% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_scale.R
\name{ts_scale}
\alias{ts_scale}
\title{Normalized Time Series}
\usage{
ts_scale(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{ts_boxable time series}

\item{center}{logical}

\item{scale}{logical}
}
\description{
Normalized Time Series
}
\examples{
ts_scale((ts_c(airmiles, co2, JohnsonJohnson, discoveries)))
ts_scale(ts_c(AirPassengers, DAX = EuStockMarkets[, 'DAX']))
}
