% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_select.R
\name{ts_select}
\alias{ts_select}
\alias{ts_select.ts}
\alias{ts_select.xts}
\alias{ts_select.data.frame}
\alias{ts_select.data.table}
\alias{ts_select.tbl}
\title{Set and Variables from Multiple Time Series}
\usage{
ts_select(x, vars, ...)

\method{ts_select}{ts}(x, vars, ...)

\method{ts_select}{xts}(x, vars, ...)

\method{ts_select}{data.frame}(x, vars, ...)

\method{ts_select}{data.table}(x, vars, ...)

\method{ts_select}{tbl}(x, vars, ...)
}
\arguments{
\item{x}{a time series object, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}

\item{vars}{character, time series to be selectd}

\item{...}{arguments passed to methods}
}
\description{
Set and Variables from Multiple Time Series
}
\examples{

all.equal(ts_ts(mdeaths), ts_select(ts_ts(ts_c(mdeaths, fdeaths)), 'mdeaths'))
all.equal(ts_xts(mdeaths), 
          ts_select(ts_xts(ts_c(mdeaths, fdeaths)), 'mdeaths'), 
          check.attributes = FALSE)
all.equal(ts_df(mdeaths), ts_select(ts_df(ts_c(mdeaths, fdeaths)), 'mdeaths'))

\dontrun{
library(data.table)  # if you want to use the 'data.table' methods
all.equal(ts_dt(mdeaths), ts_select(ts_dt(ts_c(mdeaths, fdeaths)), 'mdeaths'))
}

}
