% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_shift.R
\name{ts_shift}
\alias{ts_shift}
\title{Shift Time Stamp}
\usage{
ts_shift(x, by = NULL)
}
\arguments{
\item{x}{ts-boxable time series, an object of class \code{ts}, \code{xts}, \code{data.frame}, \code{data.table}, or \code{tibble}.}

\item{by}{passed on to \code{\link[base:seq.Date]{base::seq.Date()}}.
Argument can be specified in several ways:
- A number, taken to be in days.
- A object of class difftime
- A character string, containing one of "day", "week", "month", "quarter"
or "year". This can optionally be preceded by a (positive or negative)
integer and a space, or followed by "s".}
}
\value{
a ts-boxable time series, with the same class as the input. If time
stamp shifting causes the object to be irregular, a data frame is returned.
}
\description{
Shift time stamps in ts-boxable time series. Unlike
\code{ts_lag} this does not ensure regularity and also works with irregular
series.
}
\examples{
ts_shift(AirPassengers, "1 month")#' 
ts_shift(AirPassengers, 3)

}
\seealso{
\code{\link[=ts_lag]{ts_lag()}}, for shifting regular series. \code{\link[=date_shift]{date_shift()}}, for
shifting \code{Date} vectors.
}
