% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_span.R
\name{ts_span}
\alias{ts_span}
\alias{ts_span}
\alias{ts_start}
\alias{ts_span}
\alias{ts_end}
\title{Time Span}
\usage{
ts_span(x, start = NULL, end = NULL, template = NULL)

ts_start(x)

ts_end(x)
}
\arguments{
\item{x}{ts-boxable time series, an object of class \code{ts}, \code{xts}, \code{data.frame}, \code{data.table}, or \code{tibble}.}

\item{start}{start date, character string, \code{Date} or \code{POSIXct}}

\item{end}{end date, character string, \code{Date} or \code{POSIXct}.}

\item{template}{ts-boxable time series, an object of class \code{ts}, \code{xts},
\code{data.frame}, \code{data.table}, or \code{tibble}. If provided, \code{from} and \code{to}
will be extracted from the object.}
}
\value{
a ts-boxable time series, with the same class as the input.
}
\description{
Filter series for time span
}
\details{
All date and times, when entered as charachter strings, are processed by
\code{anytime::anydate()} or \code{anytime::anytime()}. Thus a wide range of inputs are
possible. See examples.
}
\examples{

# use 'anytime' shortcuts
ts_span(mdeaths, start = "1979")       # shortcut for 1979-01-01

ts_span(mdeaths, start = "1979-4")     # shortcut for 1979-04-01

ts_span(mdeaths, start = "197904")     # shortcut for 1979-04-01

# it's fine to use an to date outside of series span
ts_span(mdeaths, end = "2001-01-01")

# ts_end and ts_start, together with time_shift allow flexible specification
# of relative ranges:

# latest value
ts_span(mdeaths, start = time_shift(ts_end(mdeaths), -1))

ts_plot(
  ts_span(mdeaths, start = time_shift(ts_end(mdeaths), "-3 years")),
  title = "Three years",
  subtitle = "The last three years of data"
)

ts_ggplot(
  ts_span(mdeaths, end = time_shift(ts_start(mdeaths), "28 weeks")),
  title = "28 weeks later",
  subtitle = "The first 28 weeks of data"
) + theme_tsbox() + scale_color_tsbox()

# Limit span of 'discoveries' to the same span as 'AirPassengers'
ts_span(discoveries, template = AirPassengers)
}
