% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_trend.R
\name{ts_trend}
\alias{ts_trend}
\title{Loess smoothing}
\usage{
ts_trend(x, degree = 2, span = NULL)
}
\arguments{
\item{x}{any time series object}

\item{degree}{degree of Loess smoothing}

\item{span}{smoothing parameter, if NULL, automated search}
}
\description{
Loess smoothing
}
