% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_from_data.frame.R, R/to_from_data.table.R,
%   R/to_from_tibble.R, R/to_from_ts.R, R/to_from_tslist.R, R/to_from_xts.R
\name{ts_ts}
\alias{ts_ts}
\alias{ts_data.frame}
\alias{ts_ts}
\alias{ts_df}
\alias{ts_ts}
\alias{ts_data.table}
\alias{ts_ts}
\alias{ts_dt}
\alias{ts_ts}
\alias{ts_tbl}
\alias{ts_ts}
\alias{ts_ts}
\alias{ts_tslist}
\alias{ts_ts}
\alias{ts_xts}
\title{Convert Everything to Everything}
\usage{
ts_data.frame(x)

ts_df(x)

ts_data.table(x)

ts_dt(x)

ts_tbl(x)

ts_ts(x)

ts_tslist(x)

ts_xts(x)
}
\arguments{
\item{x}{ts-boxable time series, an object of class \code{ts}, \code{xts}, \code{data.frame}, \code{data.table}, or \code{tibble}.}
}
\value{
ts-boxable time series of the desired class, an object of class \code{ts}, \code{xts}, \code{data.frame}, \code{data.table}, or \code{tibble}.
}
\description{
tsbox is built around a set of converters, which convert time series
stored as \code{ts}, \code{xts}, \code{data.frame}, \code{data.table} or \code{tibble} to each
other.

Mulitple time series will be stored as a 'long' data frame (\code{data.frame},
\code{data.table} or \code{tibble}):
}
\details{
\preformatted{ts_df(ts_c(fdeaths, mdeaths))

#          id       time value
# 1   fdeaths 1974-01-01   901
# 2   fdeaths 1974-02-01   689
# 3   fdeaths 1974-03-01   827
# 4   fdeaths 1974-04-01   677
# 5   fdeaths 1974-05-01   522
# ...
# 140 mdeaths 1979-08-01   975
# 141 mdeaths 1979-09-01   940
# 142 mdeaths 1979-10-01  1081
# 143 mdeaths 1979-11-01  1294
# 144 mdeaths 1979-12-01  1341
}

The time stamp, \code{time}, indicates the beginning of a period. tsbox requires the
columns in a data frame to follow either the order:
\enumerate{
\item \strong{id} column(s)
\item \strong{time} column
\item \strong{value} column
}

\strong{or} the \strong{time} colum and the \strong{value} column to be explicitly named as \code{time} and \code{value}. If explicit names are used, the column order will be ignored.

Note that multiple id columns with arbitrary names are allowed.
}
\examples{

x.ts <- ts_c(mdeaths, fdeaths)
x.df <- ts_df(x.ts)
x.dt <- ts_dt(x.df)
\dontrun{
library(xts)
x.xts <- ts_xts(x.ts)
library(dplyr)
x.tbl <- ts_tbl(x.ts)
}

}
