% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_window.R
\name{ts_window}
\alias{ts_window}
\alias{ts_window}
\alias{ts_start}
\alias{ts_window}
\alias{ts_end}
\title{Time Windows}
\usage{
ts_window(x, start = NULL, end = NULL, template = NULL)

ts_start(x)

ts_end(x)
}
\arguments{
\item{x}{ts-boxable time series, an object of class \code{ts}, \code{xts}, \code{data.frame}, \code{data.table}, or \code{tibble}.}

\item{start}{start date, character string, \code{Date} or \code{POSIXct}}

\item{end}{end date, character string, \code{Date} or \code{POSIXct}.}

\item{template}{ts-boxable time series, an object of class \code{ts}, \code{xts},
\code{data.frame}, \code{data.table}, or \code{tibble}. If provided, \code{start} and \code{end}
will be extracted from this object.}
}
\value{
a ts-boxable time series, with the same class as the input.
}
\description{
Filter series for a time range.
}
\examples{
# It's fine to use an end date outside of series span
ts_window(mdeaths, end = "2001-01-01")

ms <- ts_window(mdeaths, end = "1976-12-01")
ts_window(ts_c(fdeaths, ms), start = ts_start(ms), end = ts_end(ms))
}
