% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsbind.R, R/tsrbind.R
\name{tsbind}
\alias{tsbind}
\alias{tsrbind}
\title{Bind any time series vertically or horizontally}
\usage{
tsbind(...)

tsrbind(...)
}
\arguments{
\item{...}{time series objects, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}
}
\description{
Bind any time series vertically or horizontally
}
\examples{

tsbind(as_df(EuStockMarkets), AirPassengers)
tsbind(EuStockMarkets, mdeaths)

# labelling:
tsbind(`International Airline Passengers` = as_xts(AirPassengers), 
       `Deaths from Lung Diseases` = ldeaths)

tsrbind(as_df(mdeaths), AirPassengers)
tsrbind(as_xts(AirPassengers), mdeaths)

}
