% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsbox-deprecated.R
\name{tsbox-deprecated}
\alias{tsbox-deprecated}
\alias{ts_start}
\alias{ts_end}
\title{Start and end of time series}
\usage{
ts_start(x)

ts_end(x)
}
\arguments{
\item{x}{ts-boxable time series, an object of class \code{ts}, \code{xts}, \code{zoo}, \code{data.frame}, \code{data.table}, \code{tbl}, \code{tbl_ts}, \code{tbl_time}, or \code{timeSeries}.}
}
\description{
Start and end of time series
}
