% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsggplot.R
\name{tsggsave}
\alias{tsggsave}
\title{ggsave, optimized for time series}
\usage{
tsggsave(filename = "myfig.pdf", width = 10, height = 5, device = "pdf",
  ..., open = TRUE)
}
\arguments{
\item{filename}{filename}

\item{width}{width}

\item{height}{height}

\item{device}{device}

\item{...}{aruments passed to ggsave}

\item{open}{should the graph be opened?}
}
\description{
ggsave, optimized for time series
}
\examples{
\dontrun{
tsggplot(AirPassengers)
tssave()
}
}
