% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsnames.R
\name{tsnames}
\alias{tsnames}
\alias{tsnames.ts}
\alias{tsnames.xts}
\alias{tsnames.data.frame}
\alias{tsnames.data.table}
\alias{settsnames}
\alias{settsnames.ts}
\alias{settsnames.xts}
\alias{settsnames.data.frame}
\alias{settsnames.data.table}
\title{Set and Extract Names of Time Series}
\usage{
tsnames(x, ...)

\method{tsnames}{ts}(x, ...)

\method{tsnames}{xts}(x, ...)

\method{tsnames}{data.frame}(x, ...)

\method{tsnames}{data.table}(x, ...)

settsnames(x, value)

\method{settsnames}{ts}(x, value)

\method{settsnames}{xts}(x, value)

\method{settsnames}{data.frame}(x, value)

\method{settsnames}{data.table}(x, value)
}
\arguments{
\item{x}{a time series object, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}

\item{...}{additional arguments, passed to methods}

\item{value}{new names}
}
\description{
Set and Extract Names of Time Series
}
\examples{

x.ts <- tsbind(mdeaths, fdeaths) 
x.xts <- as_xts(x.ts)
x.df <- as_df(x.xts)

tsnames(x.ts)
tsnames(x.xts)
tsnames(x.df)

\dontrun{
library(data.table)  # if you want to use the 'data.table' methods
x.dt <- as_dt(x.df)
tsnames(x.dt)
}

}
