% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tspc.R, R/tsscale.R, R/tstrend.R, R/tswin.R
\name{tspc}
\alias{tspc}
\alias{tspc.ts}
\alias{tspc.xts}
\alias{tspc.data.frame}
\alias{tspc.data.table}
\alias{tspcy}
\alias{tspcy.ts}
\alias{tspcy.xts}
\alias{tspcy.data.frame}
\alias{tspcy.data.table}
\alias{tsscale}
\alias{tsscale.xts}
\alias{tsscale.ts}
\alias{tsscale.data.frame}
\alias{tsscale.data.table}
\alias{tstrend}
\alias{tstrend.xts}
\alias{tstrend.ts}
\alias{tstrend.data.frame}
\alias{tstrend.data.table}
\alias{tswin}
\alias{tswin.xts}
\alias{tswin.ts}
\alias{tswin.data.frame}
\alias{tswin.data.table}
\title{Utility Functions for Time Series}
\usage{
tspc(x, ...)

\method{tspc}{ts}(x, ...)

\method{tspc}{xts}(x, ...)

\method{tspc}{data.frame}(x, ...)

\method{tspc}{data.table}(x, ...)

tspcy(x, ...)

\method{tspcy}{ts}(x, ...)

\method{tspcy}{xts}(x, ...)

\method{tspcy}{data.frame}(x, ...)

\method{tspcy}{data.table}(x, ...)

tsscale(x, ...)

\method{tsscale}{xts}(x, ...)

\method{tsscale}{ts}(x, ...)

\method{tsscale}{data.frame}(x, ...)

\method{tsscale}{data.table}(x, ...)

tstrend(x, ...)

\method{tstrend}{xts}(x, degree = 2, span = NULL, ...)

\method{tstrend}{ts}(x, ...)

\method{tstrend}{data.frame}(x, ...)

\method{tstrend}{data.table}(x, ...)

tswin(x, ...)

\method{tswin}{xts}(x, start = NULL, end = NULL, ...)

\method{tswin}{ts}(x, ...)

\method{tswin}{data.frame}(x, ...)

\method{tswin}{data.table}(x, ...)
}
\arguments{
\item{x}{a time series object, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}

\item{...}{additional arguments, passed to methods}

\item{degree}{see ?loess}

\item{span}{see ?loess}

\item{start}{start date, string, Date or POSIXct}

\item{end}{end date, string, Date or POSIXct}
}
\description{
Utility Functions for Time Series
}
\examples{
library(tsbox)

x.ts <- tsbind(mdeaths, fdeaths) 
x.xts <- as_xts(x.ts)
x.df <- as_df(x.xts)

tsscale(x.ts)
tsscale(x.xts)
tsscale(x.df)
tstrend(x.ts)
tstrend(x.xts)
tstrend(x.df)

tspc(x.ts)
tspc(x.xts)
tspc(x.df)

tspcy(x.ts)
tspcy(x.xts)
tspcy(x.df)

\dontrun{
library(data.table)  # if you want to use the 'data.table' methods
x.dt <- as_dt(x.df)
tsscale(x.dt)
tstrend(x.dt)
tspc(x.dt)
tspcy(x.dt)
}

}
