% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{tsplot}
\alias{tsplot}
\title{Beautiful Graphics}
\usage{
tsplot(x, title = NULL, subtitle = NULL, ...)
}
\description{
Beautiful Graphics
}
\examples{
library(mytools)
tsplot(AirPassengers) + ggtitle("Airline passengers", subtitle = "The classic Box & Jenkins airline data")
tsplot(cbind(total = ldeaths, female = fdeaths, male = mdeaths))

library(Quandl)
tsplot(Quandl("FRED/GDPMC1", "xts"))
ggsave("myfig.pdf", width = 8, height = 5)
}
