% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{tssave}
\alias{tssave}
\title{ggsave, optimized for time series}
\usage{
tssave(filename = "myfig.pdf", width = 7, height = 4, device = "pdf",
  ..., open = TRUE)
}
\arguments{
\item{filename}{filename}

\item{width}{width}

\item{height}{height}

\item{device}{device}

\item{...}{aruments passed to ggsave}

\item{open}{should the graph be opened?}
}
\description{
ggsave, optimized for time series
}
\examples{
\dontrun{
tsplot(AirPassengers)
tssave()
}
}
