% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsselect.R
\name{tsselect}
\alias{tsselect}
\alias{tsselect.ts}
\alias{tsselect.xts}
\alias{tsselect.data.frame}
\alias{tsselect.data.table}
\title{Set and Variables from Multiple Time Series}
\usage{
tsselect(x, var)

\method{tsselect}{ts}(x, var)

\method{tsselect}{xts}(x, var)

\method{tsselect}{data.frame}(x, var)

\method{tsselect}{data.table}(x, var)
}
\arguments{
\item{x}{a time series object, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}

\item{var}{character, time series to be selectd}
}
\description{
Set and Variables from Multiple Time Series
}
\examples{

all.equal(as_ts(mdeaths), tsselect(as_ts(tsbind(mdeaths, fdeaths)), 'mdeaths'))
all.equal(as_xts(mdeaths), 
          tsselect(as_xts(tsbind(mdeaths, fdeaths)), 'mdeaths'), 
          check.attributes = FALSE)
all.equal(as_df(mdeaths), tsselect(as_df(tsbind(mdeaths, fdeaths)), 'mdeaths'))

\dontrun{
library(data.table)  # if you want to use the 'data.table' methods
all.equal(as_dt(mdeaths), tsselect(as_dt(tsbind(mdeaths, fdeaths)), 'mdeaths'))
}

}
