% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_from_all.R
\name{copy_class}
\alias{copy_class}
\title{Reclass Time Series}
\usage{
copy_class(z, x, preserve.mode = TRUE, preserve.names = TRUE)
}
\arguments{
\item{z}{series to reclass}

\item{x}{template series}

\item{preserve.mode}{should the mode the time column be preserved (data frame only)}

\item{preserve.names}{should the name of the time column be preserved (data frame only)}
}
\description{
Copies class attributes from an existing ts-boxable series. Mainly used
internally.
}
\details{
Inspired by \code{xts::reclass}, which does something similar.
}
