% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_utils.R
\name{date_shift}
\alias{date_shift}
\alias{date_shift}
\alias{date_month}
\alias{date_shift}
\alias{date_quarter}
\alias{date_shift}
\alias{date_year}
\title{Manipulating Dates}
\usage{
date_shift(x, by = NULL)

date_month(x)

date_quarter(x)

date_year(x)
}
\arguments{
\item{x}{\code{Date} or \code{POSIXct}. If \code{POSIXct}, it is converted into \code{Date}.}

\item{by}{passed on to \code{\link[=base:seq.Date]{base:seq.Date()}}. Argument can be specified in several ways:
- A number, taken to be in days.
- A object of class difftime
- A character string, containing one of "day", "week", "month", "quarter" or "year". This can optionally be preceded by a (positive or negative) integer and a space, or followed by "s".}
}
\value{
an object of class \code{Date}
}
\description{
\code{date_shift} adds days, weeks, months, quarters or years to dates.
\code{date_month}, \code{date_quarter} and \code{date_year} return the first day of the
period and are useful for customized aggregation of data frames. For standard
aggregation, use \code{\link[=ts_frequency]{ts_frequency()}}.
}
\examples{
date_month(ts_df(AirPassengers)$time)
date_quarter(ts_df(AirPassengers)$time)

date_shift(ts_df(AirPassengers)$time, 14)
date_shift(ts_df(AirPassengers)$time, "7 week")
date_shift(ts_df(AirPassengers)$time, "-1 month")

}
\seealso{
\code{\link[=ts_frequency]{ts_frequency()}} for standard aggregation
}
