% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsggplot.R
\name{theme_ts}
\alias{theme_ts}
\alias{ts_colors}
\alias{scale_color_ts}
\alias{scale_fill_ts}
\title{ggplot Theme for tsbox}
\usage{
theme_ts(base_family = getOption("ts_font", ""), base_size = 12)

ts_colors()

scale_color_ts(...)

scale_fill_ts(...)
}
\arguments{
\item{base_family}{base font family}

\item{base_size}{base font size}

\item{...}{aruments passed to subfunctions}
}
\description{
ggplot Theme for tsbox
}
\examples{
library(tsbox)
df <- ts_df(ts_cbind(total = ldeaths, female = fdeaths, male = mdeaths))
 \dontrun{
ggplot(df, aes(x = Index, y = Value, color = Series)) + 
  geom_line() +
  ggtitle('Deaths by lung diseases', subtitle = 'United Kindom, per year') + 
  theme_ts() + 
  scale_color_ts() 

ggsave("myfig.pdf", width = 8, height = 5)
browseURL("myfig.pdf")
}
}
