% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_ggplot.R
\name{theme_ts}
\alias{theme_ts}
\alias{colors_tsbox}
\alias{scale_color_tsbox}
\alias{scale_fill_tsbox}
\title{ggplot Theme for tsbox}
\usage{
theme_ts(base_family = getOption("ts_font", ""), base_size = 12)

colors_tsbox()

scale_color_tsbox(...)

scale_fill_tsbox(...)
}
\arguments{
\item{base_family}{base font family}

\item{base_size}{base font size}

\item{...}{aruments passed to subfunctions}
}
\description{
ggplot Theme for tsbox
}
\examples{
library(tsbox)
df <- ts_df(ts_bind(total = ldeaths, female = fdeaths, male = mdeaths))
 \dontrun{
ggplot(df, aes(x = Index, y = Value, color = Series)) + 
  geom_line() +
  ggtitle('Deaths by lung diseases', subtitle = 'United Kindom, per year') + 
  theme_ts() + 
  scale_color_tsbox() 

ggsave("myfig.pdf", width = 8, height = 5)
browseURL("myfig.pdf")
}
}
