% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_ggplot.R
\name{theme_ts}
\alias{theme_ts}
\alias{colors_tsbox}
\alias{scale_color_tsbox}
\alias{scale_fill_tsbox}
\title{ggplot Theme for tsbox}
\usage{
theme_ts(base_family = getOption("ts_font", ""), base_size = 12)

colors_tsbox()

scale_color_tsbox(...)

scale_fill_tsbox(...)
}
\arguments{
\item{base_family}{base font family}

\item{base_size}{base font size}

\item{...}{aruments passed to subfunctions}
}
\description{
ggplot Theme for tsbox
}
