% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_ggplot.R
\name{theme_tsbox}
\alias{theme_tsbox}
\alias{colors_tsbox}
\alias{scale_color_tsbox}
\alias{scale_fill_tsbox}
\title{ggplot Theme for tsbox}
\usage{
theme_tsbox(base_family = getOption("ts_font", ""), base_size = 12)

colors_tsbox()

scale_color_tsbox(...)

scale_fill_tsbox(...)
}
\arguments{
\item{base_family}{base font family}

\item{base_size}{base font size}

\item{...}{aruments passed to subfunctions}
}
\description{
Using same style as \code{ts_plot()}.
}
\examples{

df <- ts_df(ts_c(total = ldeaths, female = fdeaths, male = mdeaths))
# standard ggplot
library(ggplot2)
ggplot(df, aes(x = time, y = value, color = var)) + 
  geom_line()

# a quick plot for time series
p <- ts_ggplot(AirPassengers, mdeaths)

# using same style as base graphic ts_plot()
p +
  theme_tsbox() + 
  scale_color_tsbox()

}
