% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_.R, R/ts_scale.R
\name{ts_}
\alias{ts_}
\alias{ts_diff}
\alias{ts_}
\alias{ts_scale}
\title{Universal Constructor for ts Functions}
\usage{
ts_(f, postproc = function(x) x, class = "ts", vectorize = FALSE, ...)

ts_diff(x, ...)

ts_scale(x, ...)
}
\arguments{
\item{f}{function, to be made available to all time series classes}

\item{postproc}{a function applied to the return value of \code{f()}. E.g.
\code{predict()}. (See examples)}

\item{class}{class that the function uses as its first argument}

\item{vectorize}{should the function be vectorized? (not yet implemented)}

\item{...}{arguments passed to underlying functions.}

\item{x}{time series object, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}
}
\description{
Universal Constructor for ts Functions
}
\examples{
ts_(rowSums)(ts_c(mdeaths, fdeaths))
ts_plot(mean = ts_(rowMeans)(ts_c(mdeaths, fdeaths)), mdeaths, fdeaths)
ts_(prcomp, predict)(ts_c(mdeaths, fdeaths))
ts_(prcomp, predict, scale = TRUE)(ts_c(mdeaths, fdeaths))
}
