% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_.R
\name{ts_}
\alias{ts_}
\title{universal constructor for ts functions}
\usage{
ts_(FUN, class = "ts", multiple = TRUE, suggested = NULL)
}
\arguments{
\item{FUN}{function, to be made available to all time series classes}

\item{class}{class that the function uses as its first argument}

\item{multiple}{can the function handle multiple series. If set to false, the
wrapper will loop through each series.}

\item{packages}{that are required for the functionality.}
}
\description{
universal constructor for ts functions
}
