% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_.R
\name{ts_}
\alias{ts_}
\alias{tsdiff}
\alias{tswindow}
\alias{tslag}
\alias{tscycle}
\alias{tsforecast}
\alias{tsseas}
\alias{tsprcomp}
\title{universal constructor for ts functions}
\usage{
ts_(FUN, class = "ts", multiple = TRUE, suggested = NULL)

tsdiff(x, ...)

tswindow(x, ...)

tslag(x, ...)

tscycle(x, ...)

tsforecast(x, ...)

tsseas(x, ...)

tsprcomp(x, ...)
}
\arguments{
\item{FUN}{function, to be made available to all time series classes}

\item{class}{class that the function uses as its first argument}

\item{multiple}{can the function handle multiple series. If set to false, the
wrapper will loop through each series.}

\item{suggested}{packages that are required for the functionality.}

\item{...}{arguments passed to underlying functions.}

\item{n}{how many princial components should be extracted}
}
\description{
universal constructor for ts functions
}
\examples{
tsplot(
    tsbind(AirPassengers, mdeaths),
    tsforecast(tsbind(AirPassengers, mdeaths))
)

}
