% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_.R
\docType{data}
\name{ts_}
\alias{ts_}
\alias{ts_diff}
\alias{ts_scale}
\alias{ts_windowdow}
\alias{ts_lag}
\alias{ts_cycle}
\alias{ts_forecast}
\alias{ts_forecast.auto.arima}
\alias{ts_seas}
\alias{ts_prcomp}
\alias{ts_residlm}
\alias{ts_predictlm}
\title{universal constructor for ts functions}
\format{An object of class \code{call} of length 4.}
\usage{
ts_(FUN, class = "ts", multiple = TRUE, suggested = NULL,
  ensure.names = TRUE)

ts_diff

ts_scale

ts_windowdow

ts_lag

ts_cycle

ts_forecast

ts_forecast.auto.arima

ts_seas

ts_prcomp

ts_residlm

ts_predictlm
}
\arguments{
\item{FUN}{function, to be made available to all time series classes}

\item{class}{class that the function uses as its first argument}

\item{multiple}{can the function handle multiple series. If set to false, the
wrapper will loop through each series.}

\item{suggested}{packages that are required for the functionality.}

\item{ensure.names}{logical, whether the series in the output should have the same names as the input}

\item{...}{arguments passed to underlying functions.}

\item{x}{s time series object, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}
}
\description{
universal constructor for ts functions
}
\examples{
ts_plot(
    ts_bind(AirPassengers, mdeaths),
    ts_forecast(ts_bind(AirPassengers, mdeaths))
)

}
\keyword{datasets}
