% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_.R
\name{ts_}
\alias{ts_}
\alias{load_suggested}
\alias{ts_}
\title{Universal Constructor for ts Functions}
\usage{
load_suggested(pkg)

ts_(fun, class = "ts", vectorize = FALSE, reclass = TRUE)
}
\arguments{
\item{pkg}{external package, to be suggested (automatically added by \code{ts_})
\code{predict()}. (See examples)}

\item{fun}{function, to be made available to all time series classes}

\item{class}{class that the function uses as its first argument}

\item{vectorize}{should the function be vectorized? (not yet implemented)}

\item{reclass}{logical, should the new function return the same same
ts-boxable output as imputed?}
}
\description{
Universal Constructor for ts Functions
}
\examples{
ts_(rowSums)(ts_c(mdeaths, fdeaths))
ts_plot(mean = ts_(rowMeans)(ts_c(mdeaths, fdeaths)), mdeaths, fdeaths)
ts_(function(x) predict(prcomp(x)))(ts_c(mdeaths, fdeaths))
ts_(function(x) predict(prcomp(x, scale = TRUE)))(ts_c(mdeaths, fdeaths))
ts_(dygraphs::dygraph, class = "xts")
}
