% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_.R
\name{ts_}
\alias{ts_}
\title{Universal Constructor for ts Functions}
\usage{
ts_(f, class = "ts", vectorize = FALSE)
}
\arguments{
\item{f}{function, to be made available to all time series classes}

\item{class}{class that the function uses as its first argument}

\item{vectorize}{should the function be vectorized? (not yet implemented)
\code{predict()}. (See examples)}
}
\description{
Universal Constructor for ts Functions
}
\examples{
ts_(rowSums)(ts_c(mdeaths, fdeaths))
ts_plot(mean = ts_(rowMeans)(ts_c(mdeaths, fdeaths)), mdeaths, fdeaths)
ts_(function(x) predict(prcomp(x)))(ts_c(mdeaths, fdeaths))
ts_(function(x) predict(prcomp(x, scale = TRUE)))(ts_c(mdeaths, fdeaths))
}
