% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_.R
\name{ts_}
\alias{ts_}
\alias{ts_diff}
\alias{ts_lag}
\alias{ts_cycle}
\alias{ts_seas_final}
\alias{ts_prcomp}
\title{universal constructor for ts functions}
\usage{
ts_(FUN, specific.class = "ts", multi.series = TRUE,
  suggested.packages = NULL, ensure.names = TRUE)

ts_diff(x, ...)

ts_lag(x, ...)

ts_cycle(x, ...)

ts_seas_final(x, ...)

ts_prcomp(x, ...)
}
\arguments{
\item{FUN}{function, to be made available to all time series classes}

\item{multi.series}{can the function handle multiple series. If set to false, the
wrapper will loop through each series.}

\item{suggested.packages}{packages that are required for the functionality.}

\item{ensure.names}{logical, whether the series in the output should have the same names as the input}

\item{x}{s time series object, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}

\item{...}{arguments passed to underlying functions.}

\item{class}{class that the function uses as its first argument}
}
\description{
universal constructor for ts functions
}
\examples{
ts_plot(
    ts_c(AirPassengers, mdeaths),
    ts_forecast(ts_c(AirPassengers, mdeaths))
)

}
