% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_window.R
\name{ts_align}
\alias{ts_align}
\title{Aligning Time Series}
\usage{
ts_align(x, with)
}
\arguments{
\item{x}{any time series object}

\item{with}{character string, the variable the remaining data is aligned with}
}
\description{
Aligning Time Series
}
