% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_union.R
\name{ts_align}
\alias{ts_align}
\alias{ts_align}
\alias{ts_union}
\title{Align Time Series}
\usage{
ts_align(x, fill = NA)

ts_union(x, fill = NA)
}
\arguments{
\item{x}{ts-boxable time series, an object of class \code{ts}, \code{xts}, \code{data.frame}, \code{data.table}, or \code{tibble}.}

\item{fill}{missign value specifier}
}
\value{
a ts-boxable time series, with the same class as the input.
}
\description{
\code{ts_align} makes sure all time series cover the same time span and have the
same regular frequency. \code{ts_union} only ensures that series have the same
time stamps and also works with irregular series. For \code{ts} objects, the two
functions have the same effect.
}
\examples{
ts_align(ts_df(ts_c(mdeaths, fdeaths = ts_window(fdeaths, end = "1977-01-01"))))
}
