% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_arithmetic.R
\name{ts_arithmetic}
\alias{ts_arithmetic}
\alias{\%ts+\%}
\alias{\%ts-\%}
\alias{\%ts*\%}
\alias{\%ts/\%}
\title{Arithmetic Operators for ts-boxable objects}
\usage{
e1 \%ts+\% e2

e1 \%ts-\% e2

e1 \%ts*\% e2

e1 \%ts/\% e2
}
\arguments{
\item{e1}{ts-boxable time series, an object of class \code{ts}, \code{xts}, \code{zoo},
\code{data.frame}, \code{data.table}, \code{tbl}, \code{tbl_ts}, \code{tbl_time}, \code{tis}, \code{irts} or
\code{timeSeries}.}

\item{e2}{ts-boxable time series, an object of class \code{ts}, \code{xts}, \code{zoo},
\code{data.frame}, \code{data.table}, \code{tbl}, \code{tbl_ts}, \code{tbl_time}, \code{tis}, \code{irts} or
\code{timeSeries}.}
}
\value{
a ts-boxable time series, with the same class as the left input.
}
\description{
Arithmetic Operators for ts-boxable objects
}
\examples{
head(fdeaths - mdeaths)
head(fdeaths \%ts-\% mdeaths)
head(ts_df(fdeaths) \%ts-\% mdeaths)
}
