% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsbind.R, R/tsrbind.R
\name{ts_bind}
\alias{ts_bind}
\alias{ts_rbind}
\title{Bind any time series vertically or horizontally}
\usage{
ts_bind(...)

ts_rbind(...)
}
\arguments{
\item{...}{time series objects, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}
}
\description{
Bind any time series vertically or horizontally
}
\examples{

ts_bind(ts_df(EuStockMarkets), AirPassengers)
ts_bind(EuStockMarkets, mdeaths)

# labelling:
ts_bind(`International Airline Passengers` = ts_xts(AirPassengers), 
       `Deaths from Lung Diseases` = ldeaths)

ts_rbind(ts_df(mdeaths), AirPassengers)
ts_rbind(ts_xts(AirPassengers), mdeaths)

}
