% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_bind.R, R/ts_chain.R
\name{ts_bind}
\alias{ts_bind}
\alias{ts_bind}
\alias{ts_chain}
\title{Combine to Single Time Series}
\usage{
ts_bind(...)

ts_chain(...)
}
\arguments{
\item{...}{one or several tsboxable time series}
}
\description{
Combine time series to a new, single time series. \code{ts_bind} combines time
series as they are, \code{ts_chain} chaines them together, using percentage change
rates.

Bind to or several tsboxable objects
}
