% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_.R
\name{ts_boxable}
\alias{ts_boxable}
\title{Test if object is a valid time series}
\usage{
ts_boxable(x)
}
\arguments{
\item{x}{time series object, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}
}
\description{
Test if object is a valid time series
}
