% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_from_all.R
\name{ts_boxable}
\alias{ts_boxable}
\title{Test if an Object is ts-Boxable}
\usage{
ts_boxable(x)
}
\arguments{
\item{x}{ts-boxable time series, an object of class \code{ts}, \code{xts}, \code{zoo}, \code{data.frame}, \code{data.table}, \code{tbl}, \code{tbl_ts}, \code{tbl_time}, or \code{timeSeries}.}
}
\value{
logical, either \code{TRUE} or \code{FALSE}
}
\description{
Mainly used internally.
}
\examples{
ts_boxable(AirPassengers)
ts_boxable(lm)
}
