% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_c.R
\name{ts_c}
\alias{ts_c}
\title{Bind any time series vertically or horizontally}
\usage{
ts_c(...)
}
\arguments{
\item{...}{time series objects, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}
}
\description{
Bind any time series vertically or horizontally
}
\examples{

ts_c(ts_df(EuStockMarkets), AirPassengers)
a <- ts_c(EuStockMarkets, mdeaths)

# labelling:
ts_c(`International Airline Passengers` = ts_xts(AirPassengers),
       `Deaths from Lung Diseases` = ldeaths)

ts_c(a = mdeaths, AirPassengers)
ts_bind(ts_xts(AirPassengers), mdeaths)
ts_c(ts_dt(EuStockMarkets), AirPassengers)
ts_bind(ts_dt(mdeaths), AirPassengers)

}
