% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_c.R, R/ts_chain.R, R/ts_rbind.R
\name{ts_c}
\alias{ts_c}
\alias{ts_chain}
\alias{ts_rbind}
\title{Bind any time series vertically or horizontally}
\usage{
ts_c(...)

ts_chain(...)

ts_rbind(...)
}
\arguments{
\item{...}{time series objects, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}
}
\description{
Bind any time series vertically or horizontally
}
\examples{

ts_c(ts_df(EuStockMarkets), AirPassengers)
ts_c(EuStockMarkets, mdeaths)

# labelling:
ts_c(`International Airline Passengers` = ts_xts(AirPassengers), 
       `Deaths from Lung Diseases` = ldeaths)

ts_c(a = ts_df(mdeaths), AirPassengers)
ts_rbind(ts_xts(AirPassengers), mdeaths)
ts_c(ts_dt(EuStockMarkets), AirPassengers)
ts_select(ts_c(mdeaths, austres, AirPassengers, DAX = EuStockMarkets[,'DAX']), 'mdeaths')

}
