% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts.R
\name{ts_dts}
\alias{ts_dts}
\title{Internal Time Series Class}
\usage{
ts_dts(x)
}
\arguments{
\item{x}{ts-boxable time series, an object of class \code{ts}, \code{xts}, \code{zoo}, \code{data.frame}, \code{data.table}, \code{tbl}, \code{tbl_ts}, \code{tbl_time}, \code{tis}, \code{irts} or \code{timeSeries}.}
}
\description{
Internal Time Series Class
}
