% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_pc.R
\name{ts_lag}
\alias{ts_lag}
\alias{ts_lag}
\alias{ts_pc}
\title{Lags, Differences, Percentage Change Rates}
\usage{
ts_lag(x, lag = 1, fill = NA)

ts_pc(x, ...)
}
\arguments{
\item{x}{a ts_boxable object}

\item{lag}{defined as in dplyr, opposite to R base}

\item{fill}{how to fill missing values}
}
\description{
Lags, Differences, Percentage Change Rates
}
\examples{
ts_lag(ts_c(fdeaths, mdeaths))
ts_diff(ts_c(fdeaths, mdeaths))
ts_pc(ts_c(fdeaths, mdeaths))
}
