% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_pc.R
\name{ts_pc}
\alias{ts_pc}
\alias{ts_pc}
\alias{ts_diff}
\alias{ts_pc}
\alias{ts_pcy}
\alias{ts_pc}
\alias{ts_diffy}
\title{First Differences and Percentage Change Rates}
\usage{
ts_pc(x)

ts_diff(x)

ts_pcy(x)

ts_diffy(x)
}
\arguments{
\item{x}{ts-boxable time series, an object of class \code{ts}, \code{xts}, \code{zoo}, \code{data.frame}, \code{data.table}, \code{tibble}, \code{tsibble} or \code{timeSeries}.}
}
\value{
a ts-boxable time series, with the same class as the input.
}
\description{
\code{ts_pcy} and \code{ts_diffy} calculate the percentage change rate and the difference
compared to the previous period, \code{ts_pcy} and \code{ts_diffy} calculate compared to
the same period of the previous year.
}
\examples{
head(ts_diff(ts_c(fdeaths, mdeaths)))
head(ts_pc(ts_c(fdeaths, mdeaths)))
head(ts_pcy(ts_c(fdeaths, mdeaths)))
head(ts_diffy(ts_c(fdeaths, mdeaths)))
}
