% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsggplot.R, R/tsplot.R
\name{ts_ggplot}
\alias{ts_ggplot}
\alias{ts_plot}
\title{Plot Time Series}
\usage{
ts_ggplot(..., title = NULL, subtitle = NULL)

ts_plot(..., title, subtitle, ylab = "")
}
\arguments{
\item{...}{time series objects, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}

\item{title}{title (optional)}

\item{subtitle}{subtitle (optional)}

\item{ylab}{ylab (optional)}
}
\description{
Plot Time Series
}
\examples{

library(tsbox)
ts_plot(AirPassengers, title = "Airline passengers", 
       subtitle = "The classic Box & Jenkins airline data")
ts_plot(total = ldeaths, female = fdeaths, male = mdeaths)

ts_plot(ts_cbind(sunspot.month, sunspot.year, lynx))
ts_plot(ts_scale(ts_cbind(airmiles, co2, JohnsonJohnson, discoveries)))
ts_plot(EuStockMarkets)
ts_plot(sunspot.month, sunspot.year, lynx)
ts_plot(ts_scale(ts_cbind(Nile, nottem, USAccDeaths)))

ts_ggplot(AirPassengers, title = "Airline passengers", 
       subtitle = "The classic Box & Jenkins airline data")
ts_ggplot(total = ldeaths, female = fdeaths, male = mdeaths)

ts_ggplot(ts_cbind(sunspot.month, sunspot.year, lynx))
ts_ggplot(ts_scale(ts_cbind(airmiles, co2, JohnsonJohnson, discoveries)))
ts_ggplot(EuStockMarkets)
ts_ggplot(sunspot.month, sunspot.year, lynx)
ts_ggplot(ts_scale(ts_cbind(Nile, nottem, USAccDeaths)))
\dontrun{
library(Quandl)
ts_ggplot(Quandl("FRED/GDPMC1", "xts"), title = "US GDP")

library(dataseries)
dta <- ds(c("GDP.PBRTT.A.R", "CCI.CCIIR"), "xts")
ts_ggplot(ts_scale(ts_window(ts_cbind(`GDP Growth` = ts_pc(dta[, 'GDP.PBRTT.A.R']), 
                            `Consumer Sentiment Index` = dta[, 'CCI.CCIIR']), 
                     start = "1995-01-01")),
       title = "GDP and Consumer Sentiment",
       subtitle = "normalized values")

}
}
