% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_plot.R
\name{ts_save}
\alias{ts_save}
\title{Save Previous Plot}
\usage{
ts_save(filename = tempfile(fileext = ".pdf"), width = 10,
  height = 5, device = NULL, open = TRUE)
}
\arguments{
\item{filename}{filename}

\item{width}{width}

\item{height}{height}

\item{device}{device}

\item{open}{logical, should the saved plot be opened?}
}
\description{
Save Previous Plot
}
