% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_scale.R
\name{ts_scale}
\alias{ts_scale}
\title{Normalized Time Series}
\usage{
ts_scale(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{ts_boxable time series}

\item{center}{logical}

\item{scale}{logical}
}
\description{
Subtract mean and divide by standard deviation. Based on \code{\link[base:scale]{base::scale()}}.
}
\examples{
\donttest{
ts_plot(ts_scale((ts_c(airmiles, co2, JohnsonJohnson, discoveries))))
ts_plot(ts_scale(ts_c(AirPassengers, DAX = EuStockMarkets[, 'DAX'])))
}
}
