% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsselect.R
\name{ts_select}
\alias{ts_select}
\alias{ts_select.ts}
\alias{ts_select.xts}
\alias{ts_select.data.frame}
\alias{ts_select.data.table}
\alias{ts_select.tbl}
\title{Set and Variables from Multiple Time Series}
\usage{
ts_select(x, var, ...)

\method{ts_select}{ts}(x, var, ...)

\method{ts_select}{xts}(x, var, ...)

\method{ts_select}{data.frame}(x, var, var.name = getOption("tsbox.var.name",
  "var"), ...)

\method{ts_select}{data.table}(x, var, var.name = getOption("tsbox.var.name",
  "var"), ...)

\method{ts_select}{tbl}(x, var, var.name = getOption("tsbox.var.name", "var"),
  ...)
}
\arguments{
\item{x}{a time series object, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}

\item{var}{character, time series to be selectd}

\item{...}{arguments passed to methods}

\item{var.name}{name of the column that contains \code{var}}
}
\description{
Set and Variables from Multiple Time Series
}
\examples{

all.equal(ts_ts(mdeaths), ts_select(ts_ts(ts_cbind(mdeaths, fdeaths)), 'mdeaths'))
all.equal(ts_xts(mdeaths), 
          ts_select(ts_xts(ts_cbind(mdeaths, fdeaths)), 'mdeaths'), 
          check.attributes = FALSE)
all.equal(ts_df(mdeaths), ts_select(ts_df(ts_cbind(mdeaths, fdeaths)), 'mdeaths'))

\dontrun{
library(data.table)  # if you want to use the 'data.table' methods
all.equal(ts_dt(mdeaths), ts_select(ts_dt(ts_cbind(mdeaths, fdeaths)), 'mdeaths'))
}

}
