% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_summary.R
\name{ts_summary}
\alias{ts_summary}
\title{Time Series Properties}
\usage{
ts_summary(x, spark.width = 15)
}
\arguments{
\item{x}{ts-boxable time series, an object of class \code{ts}, \code{xts}, \code{zoo}, \code{data.frame}, \code{data.table}, \code{tbl}, \code{tbl_ts}, \code{tbl_time}, or \code{timeSeries}.}

\item{spark.width}{integer, width of the spark line, in numer of characters.
Set to \code{NULL} to turn off (and speed it up).}
}
\description{
Time Series Properties
}
\examples{
ts_summary(ts_c(mdeaths, austres))
# Extracting specific properties
ts_summary(AirPassengers)$start
ts_summary(AirPassengers)$freq
ts_summary(AirPassengers)$obs
}
