% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_trend.R
\name{ts_trend}
\alias{ts_trend}
\title{Loess smoothing}
\usage{
ts_trend(x, ...)
}
\arguments{
\item{x}{any time series object}

\item{...}{arguments, passed to subfunction:
\itemize{
\item \code{degree} degree of Loess smoothing
\item \code{span} smoothing parameter, if NULL, automated search
}}
}
\description{
Loess smoothing
}
