% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_from_data.frame.R, R/to_from_data.table.R,
%   R/to_from_tibble.R, R/to_from_ts.R, R/to_from_tslist.R, R/to_from_xts.R,
%   R/ts_lag.R
\name{ts_ts}
\alias{ts_ts}
\alias{ts_data.frame}
\alias{ts_ts}
\alias{ts_df}
\alias{ts_ts}
\alias{ts_data.table}
\alias{ts_ts}
\alias{ts_dt}
\alias{ts_ts}
\alias{ts_tbl}
\alias{ts_ts}
\alias{ts_ts}
\alias{ts_tslist}
\alias{ts_ts}
\alias{ts_xts}
\alias{ts_ts}
\alias{ts_lag}
\title{Convert everything to everything}
\usage{
ts_data.frame(x)

ts_df(x)

ts_data.table(x)

ts_dt(x)

ts_tbl(x)

ts_ts(x)

ts_tslist(x)

ts_xts(x)

ts_lag(x, lag = 1, fill = NA)
}
\arguments{
\item{x}{a time series object, either \code{ts}, \code{data.frame}, \code{data.table}, \code{tibble} or \code{xts}.}

\item{lag}{defined as in dplyr, opposite to R base}

\item{fill}{how to fill missing values}
}
\description{
Convert everything to everything
}
\examples{

x.ts <- ts_c(mdeaths, fdeaths) 
x.df <- ts_df(x.ts)
x.dt <- ts_dt(x.df)
\dontrun{
library(xts)
x.xts <- ts_xts(x.ts)
library(dplyr)
x.tbl <- ts_tbl(x.ts)
}

}
