% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts.R, R/dts_data.frame.R, R/dts_data.table.R,
%   R/dts_ts.R, R/dts_xts.R
\name{ts_ts}
\alias{ts_ts}
\alias{ts_dts}
\alias{ts_ts}
\alias{ts_data.frame}
\alias{ts_ts}
\alias{ts_data.table}
\alias{ts_ts}
\alias{ts_ts}
\alias{ts_xts}
\title{Convert everything to everything}
\usage{
ts_dts(x, ...)

ts_data.frame(x, ...)

ts_data.table(x, ...)

ts_ts(x, ...)

ts_xts(x, ...)
}
\arguments{
\item{x}{a time series object, either \code{ts}, \code{data.frame}, \code{data.table}, \code{tibble} or \code{xts}.}

\item{...}{additional arguments, passed to methods}
}
\description{
Convert everything to everything
}
\examples{

x.ts <- ts_bind(mdeaths, fdeaths) 
x.df <- ts_df(x.xts)
x.dt <- ts_dt(x.xts)
\dontrun{
library(xts)
x.xts <- ts_xts(x.ts)
library(dplyr)
x.tbl <- ts_tbl(x.ts)
}

}
