% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dts.R, R/dts_data.frame.R, R/dts_data.table.R,
%   R/dts_ts.R, R/dts_xts.R, R/ts_lag.R, R/ts_scale.R
\name{ts_ts}
\alias{ts_ts}
\alias{ts_dts}
\alias{ts_ts}
\alias{ts_c.dts}
\alias{ts_ts}
\alias{ts_data.frame}
\alias{ts_ts}
\alias{ts_data.frame.dts}
\alias{ts_ts}
\alias{ts_tbl}
\alias{ts_ts}
\alias{ts_df}
\alias{ts_ts}
\alias{ts_data.table}
\alias{ts_ts}
\alias{ts_dt}
\alias{ts_ts}
\alias{ts_ts}
\alias{ts_ts.dts}
\alias{ts_ts}
\alias{ts_xts}
\alias{ts_ts}
\alias{ts_lag}
\alias{ts_ts}
\alias{ts_scale}
\title{Convert everything to everything}
\usage{
ts_dts(x, ...)

ts_c.dts(...)

ts_data.frame(x, ...)

\method{ts_data.frame}{dts}(x, ...)

ts_tbl(x, ...)

ts_df(x, ...)

ts_data.table(x, ...)

ts_dt(x, ...)

ts_ts(x, ...)

\method{ts_ts}{dts}(x, ...)

ts_xts(x, ...)

ts_lag(x, lag = 1, fill = NA)

ts_scale(x, ...)
}
\arguments{
\item{x}{a time series object, either \code{ts}, \code{data.frame}, \code{data.table}, \code{tibble} or \code{xts}.}

\item{...}{additional arguments, passed to methods}

\item{lag}{defined as in dplyr, opposite to R base}

\item{fill}{how to fill missing values}
}
\description{
Convert everything to everything
}
\examples{

x.ts <- ts_c(mdeaths, fdeaths) 
x.df <- ts_df(x.ts)
x.dt <- ts_dt(x.df)
\dontrun{
library(xts)
x.xts <- ts_xts(x.ts)
library(dplyr)
x.tbl <- ts_tbl(x.ts)
}

}
