% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_from_data.frame.R, R/to_from_data.table.R,
%   R/to_from_tibble.R, R/to_from_ts.R, R/to_from_tslist.R, R/to_from_xts.R
\name{ts_ts}
\alias{ts_ts}
\alias{ts_data.frame}
\alias{ts_ts}
\alias{ts_df}
\alias{ts_ts}
\alias{ts_data.table}
\alias{ts_ts}
\alias{ts_dt}
\alias{ts_ts}
\alias{ts_tbl}
\alias{ts_ts}
\alias{ts_ts}
\alias{ts_tslist}
\alias{ts_ts}
\alias{ts_xts}
\title{Convert Everything to Everything}
\usage{
ts_data.frame(x)

ts_df(x)

ts_data.table(x)

ts_dt(x)

ts_tbl(x)

ts_ts(x)

ts_tslist(x)

ts_xts(x)
}
\arguments{
\item{x}{ts-boxable time series, an object of class \code{ts}, \code{xts}, \code{data.frame}, \code{data.table}, or \code{tibble}.}
}
\value{
ts-boxable time series of the desired class, an object of class \code{ts}, \code{xts}, \code{data.frame}, \code{data.table}, or \code{tibble}.
}
\description{
tsbox is built around a set of converters, which convert time series
stored as \code{ts}, \code{xts}, \code{data.frame}, \code{data.table} or \code{tibble} to each
other.

Mulitple time series will be stored as a 'long' data frame (\code{data.frame},
\code{data.table} or \code{tibble}):
}
\details{
\preformatted{ts_df(ts_c(fdeaths, mdeaths))

#          id       time value
# 1   fdeaths 1974-01-01   901
# 2   fdeaths 1974-02-01   689
# 3   fdeaths 1974-03-01   827
# 4   fdeaths 1974-04-01   677
# 5   fdeaths 1974-05-01   522
# ...
# 140 mdeaths 1979-08-01   975
# 141 mdeaths 1979-09-01   940
# 142 mdeaths 1979-10-01  1081
# 143 mdeaths 1979-11-01  1294
# 144 mdeaths 1979-12-01  1341
}

The time stamp, \code{time}, indicates the beginning of a period. tsbox requires the
columns in a data frame to follow either the order:
\enumerate{
\item \strong{id} column(s)
\item \strong{time} column
\item \strong{value} column
}

\strong{or} the \strong{time} column and the \strong{value} column to be explicitly named as \code{time} and \code{value}. If explicit names are used, the column order will be ignored.

Note that multiple id columns with arbitrary names are allowed.

Whenever possible, tsbox relies on \strong{heuristic time conversion}. When a
monthly \code{"ts"} time series, e.g., \code{AirPassengers}, is converted to a data
frame, each time stamp (of class \code{"Date"}) is the first day of the month. In
most circumstances, this reflects the actual meaning of the data stored in a
\code{"ts"} object. Technically, of course, this is not correct: \code{"ts"} objects
divide time in period of equal length, while in reality, February is shorter
than January. Heuristic conversion is done for frequencies of 0.1 (decades),
1 (years), 4 (quarters) and 12 (month).

For other frequencies, e.g. 260, of \code{EuStockMarkets}, tsbox uses  \strong{exact
time conversion}. The year is divided into 260 equally long units, and time
stamp of a period will be a point in time (of class \code{"POSIXct"}).
}
\examples{

x.ts <- ts_c(mdeaths, fdeaths)
x.df <- ts_df(x.ts)
x.dt <- ts_dt(x.df)

# heuristic time conversion
ts_df(AirPassengers)

# exact time conversion
ts_df(EuStockMarkets)

# Multiple IDs
multi.id.df <- rbind(
  within(ts_df(ts_c(fdeaths, mdeaths)), type <- "level"),
  within(ts_pc(ts_df(ts_c(fdeaths, mdeaths))), type <- "pc")
)
ts_plot(multi.id.df)
ts_ts(multi.id.df)

\dontrun{
library(xts)
x.xts <- ts_xts(x.ts)
library(dplyr)
x.tbl <- ts_tbl(x.ts)
}

}
