% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ts_window}
\alias{ts_window}
\title{Deprecated}
\usage{
ts_window(x, start = NULL, end = NULL, template = NULL)
}
\arguments{
\item{x}{ts-boxable time series, an object of class \code{ts}, \code{xts}, \code{zoo}, \code{data.frame}, \code{data.table}, \code{tibble}, \code{tsibble} or \code{timeSeries}.}

\item{start}{start date, character string, \code{Date} or \code{POSIXct}}

\item{end}{end date, character string, \code{Date} or \code{POSIXct}.}

\item{template}{ts-boxable time series, an object of class \code{ts}, \code{xts},}
}
\description{
Deprecated
}
