% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_window.R
\name{ts_window}
\alias{ts_window}
\title{Agnostic Time Windows}
\usage{
ts_window(x, start = NULL, end = NULL)
}
\arguments{
\item{x}{any time series object}

\item{start}{start date, string, Date or POSIXct}

\item{end}{end date, string, Date or POSIXct}
}
\description{
Agnostic Time Windows
}
