% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsplot.R
\name{tsplot}
\alias{tsplot}
\alias{tsplot.numeric}
\alias{tsplot.ts}
\alias{tsplot.xts}
\alias{tsplot.data.frame}
\alias{tsplot.data.table}
\title{Plot Time Series}
\usage{
tsplot(..., title = NULL, subtitle = NULL)

\method{tsplot}{numeric}(..., title = NULL, subtitle = NULL)

\method{tsplot}{ts}(..., title = NULL, subtitle = NULL)

\method{tsplot}{xts}(..., title = NULL, subtitle = NULL)

\method{tsplot}{data.frame}(..., title = NULL, subtitle = NULL)

\method{tsplot}{data.table}(..., title = NULL, subtitle = NULL)
}
\arguments{
\item{...}{time series objects, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}

\item{title}{title}

\item{subtitle}{subtitle}
}
\description{
Plot Time Series
}
\examples{
\dontrun{
library(tsbox)
tsplot(AirPassengers, title = "Airline passengers", 
       subtitle = "The classic Box & Jenkins airline data")
tsplot(total = ldeaths, female = fdeaths, male = mdeaths)

tsplot(tsbind(sunspot.month, sunspot.year, lynx))
tsplot(tsscale(tsbind(airmiles, co2, JohnsonJohnson, discoveries)))
tsplot(EuStockMarkets)
tsplot(sunspot.month, sunspot.year, lynx)
tsplot(tsscale(tsbind(Nile, nottem, USAccDeaths)))

library(Quandl)
tsplot(Quandl("FRED/GDPMC1", "xts"), title = "US GDP")

library(dataseries)
dta <- ds(c("GDP.PBRTT.A.R", "CCI.CCIIR"), "xts")
tsplot(tsscale(tswin(tsbind(`GDP Growth` = tspc(dta[, 'GDP.PBRTT.A.R']), 
                            `Consumer Sentiment Index` = dta[, 'CCI.CCIIR']), 
                     start = "1995-01-01")),
       title = "GDP and Consumer Sentiment",
       subtitle = "normalized values")

}
}
