% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsggplot.R, R/tsplot.R
\name{tsggplot}
\alias{tsggplot}
\alias{tsggplot.numeric}
\alias{tsggplot.ts}
\alias{tsggplot.xts}
\alias{tsggplot.data.frame}
\alias{tsggplot.data.table}
\alias{tsplot}
\title{Plot Time Series}
\usage{
tsggplot(..., title = NULL, subtitle = NULL)

\method{tsggplot}{numeric}(..., title = NULL, subtitle = NULL)

\method{tsggplot}{ts}(..., title = NULL, subtitle = NULL)

\method{tsggplot}{xts}(..., title = NULL, subtitle = NULL)

\method{tsggplot}{data.frame}(..., title = NULL, subtitle = NULL)

\method{tsggplot}{data.table}(..., title = NULL, subtitle = NULL)

tsplot(..., title, subtitle, ylab = "",
  variable.name = getOption("tsbox.variable.name", "variable"))
}
\arguments{
\item{...}{time series objects, either \code{ts}, \code{xts}, \code{data.frame} or \code{data.table}.}

\item{title}{title (optional)}

\item{subtitle}{subtitle (optional)}

\item{ylab}{ylab (optional)}
}
\description{
Plot Time Series
}
\examples{

library(tsbox)
tsplot(AirPassengers, title = "Airline passengers", 
       subtitle = "The classic Box & Jenkins airline data")
tsplot(total = ldeaths, female = fdeaths, male = mdeaths)

tsplot(tsbind(sunspot.month, sunspot.year, lynx))
tsplot(tsscale(tsbind(airmiles, co2, JohnsonJohnson, discoveries)))
tsplot(EuStockMarkets)
tsplot(sunspot.month, sunspot.year, lynx)
tsplot(tsscale(tsbind(Nile, nottem, USAccDeaths)))

tsggplot(AirPassengers, title = "Airline passengers", 
       subtitle = "The classic Box & Jenkins airline data")
tsggplot(total = ldeaths, female = fdeaths, male = mdeaths)

tsggplot(tsbind(sunspot.month, sunspot.year, lynx))
tsggplot(tsscale(tsbind(airmiles, co2, JohnsonJohnson, discoveries)))
tsggplot(EuStockMarkets)
tsggplot(sunspot.month, sunspot.year, lynx)
tsggplot(tsscale(tsbind(Nile, nottem, USAccDeaths)))
\dontrun{
library(Quandl)
tsggplot(Quandl("FRED/GDPMC1", "xts"), title = "US GDP")

library(dataseries)
dta <- ds(c("GDP.PBRTT.A.R", "CCI.CCIIR"), "xts")
tsggplot(tsscale(tswin(tsbind(`GDP Growth` = tspc(dta[, 'GDP.PBRTT.A.R']), 
                            `Consumer Sentiment Index` = dta[, 'CCI.CCIIR']), 
                     start = "1995-01-01")),
       title = "GDP and Consumer Sentiment",
       subtitle = "normalized values")

}
}
